-FIX DRIVER MATCHING ENGINE -MODIFY AUTH -PROPER WORKING AND TESTING DONE ...EC2 INTEGRATION TESTING REMAING

+++ Backened/notification-service/src/main/java/com/fooddelivery/notification_service/dto/NotificationEvent.java
@@ -3,6 +3,8 @@ package com.fooddelivery.notification_service.dto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class NotificationEvent {
@@ -10,6 +12,6 @@ public class NotificationEvent {
     private String userEmail;
     private String userName;
     private String restaurantName;
-    private Double totalAmount;
+    private BigDecimal totalAmount;   // matches BigDecimal in both OrderPlacedEvent and PaymentCompletedEvent
     private String status;
 }
